requirements.txt
# --- Core RAG & LLM Frameworks ---
langchain==0.3.27
langchain-core==0.3.78
langchain-community==0.3.30
langchain-google-genai==2.1.12
google-ai-generativelanguage==0.7.0
sentence-transformers==2.7.0
torch==2.2.2

# --- Vector Store & Persistence ---
chromadb==0.4.18
chroma-hnswlib==0.7.3
sqlite-utils==3.35.2
SQLAlchemy==2.0.43

# --- Backend API (FastAPI) ---
fastapi==0.118.0
uvicorn==0.37.0
starlette==0.48.0
pydantic==2.11.10
python-dotenv==1.0.0

# --- Data Processing & Scientific Computing ---
pandas==2.3.3
numpy==1.26.4
Faker==37.8.0

# --- Frontend (Streamlit) ---
streamlit==1.28.1